user auth security update

+++ reviews/views.py
@@ -85,6 +85,7 @@ def comment_create(request, pk):
 def comment_delete(request, review_pk, comment_pk):
     review = Review.objects.get(pk=review_pk)
     comment = review.comment_set.get(pk=comment_pk)
-    if request.method == "POST":
-        comment.delete()
+    if request.user == comment.user:
+        if request.method == "POST":
+            comment.delete()
     return redirect("reviews:detail", review.pk)
